Use a simpler style of intrusive shared ptr.
Needs performance testing against the more complex original.

Also has a problem that needs fixing: You can instantiate an IntrusivePtr<Foo> even if Foo doesn’t inherit from the base class that makes things work. This is necessary because I want to be able to instantiate an InstrusivePtr of an incomplete type. Would be nice if was a linker error or something.